resolve warnings (style:cop:)
