+++ SixLabors.ruleset
