+++ samples/DrawWithImageSharp/DrawWithImageSharp.csproj
@@ -13,6 +13,7 @@
 
   <ItemGroup>
     <PackageReference Include="SixLabors.ImageSharp.Drawing" Version="1.0.0-beta0003" />
+    <PackageReference Include="System.Numerics.Vectors" Version="4.5.0-rc1" />
     <PackageReference Include="System.ValueTuple" Version="4.4.0" />
   </ItemGroup>
