+++ src/SixLabors.Fonts/AppliedFontStyle.cs
@@ -1,0 +1,15 @@
+﻿// Copyright (c) Six Labors and contributors.
+// Licensed under the Apache License, Version 2.0.
+
+namespace SixLabors.Fonts
+{
+    internal struct AppliedFontStyle
+    {
+        public IFontInstance Font;
+        public float PointSize;
+        public float TabWidth;
+        public int Start;
+        public int End;
+        public bool ApplyKerning;
+    }
+}
