+++ src/SixLabors.Fonts/IFontInstance.cs
@@ -8,13 +8,19 @@ namespace SixLabors.Fonts
     internal interface IFontInstance
     {
         FontDescription Description { get; }
+
         ushort EmSize { get; }
+
         int LineHeight { get; }
+
         short Ascender { get; }
+
         short Descender { get; }
+
         short LineGap { get; }
 
         GlyphInstance GetGlyph(char character);
+
         Vector2 GetOffset(GlyphInstance glyph, GlyphInstance previousGlyph);
     }
 }
