+++ src/SixLabors.Fonts/SixLabors.Fonts.csproj
@@ -35,7 +35,7 @@
     </ItemGroup>
 
     <PropertyGroup>
-        <CodeAnalysisRuleSet>..\SixLabors.ruleset</CodeAnalysisRuleSet>
+        <CodeAnalysisRuleSet>..\..\SixLabors.ruleset</CodeAnalysisRuleSet>
     </PropertyGroup>
 
     <ItemGroup>
