+++ src/SixLabors.Fonts/Tables/General/Glyphs/EmptyGlyphLoader.cs
@@ -20,12 +20,14 @@ namespace SixLabors.Fonts.Tables.General.Glyphs
         {
             if (this.loop)
             {
-                if(this.glyph == null)
+                if (this.glyph == null)
                 {
                     this.glyph = new GlyphVector(new Vector2[0], new bool[0], new ushort[0], this.fallbackEmptyBounds);
                 }
+
                 return this.glyph.Value;
             }
+
             this.loop = true;
             return table.GetGlyph(0);
         }
